move adt to adt
